--- hesai_pandar/include/hesai_pandar/input.h
@@ -1,10 +1,0 @@
-#pragma once
-#include <pandar_msgs/PandarPacket.h>
-
-
-class Input
-{
-public:
-  virtual ~Input(){};
-  virtual int getPacket(pandar_msgs::PandarPacket * pkt) = 0;
-};
