--- hesai_pandar/include/hesai_pandar/point_types.h
@@ -1,35 +1,0 @@
-#pragma once
-
-#include <pcl_ros/point_cloud.h>
-#include <pcl/point_types.h>
-
-
-struct PointXYZIR
-{
-  PCL_ADD_POINT4D;
-  float intensity;
-  uint16_t ring;
-  EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-} EIGEN_ALIGN16;
-
-struct PointXYZIRADT
-{
-  PCL_ADD_POINT4D;
-  float intensity;
-  uint16_t ring;
-  float azimuth;
-  float distance;
-  double time_stamp;
-  EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-} EIGEN_ALIGN16;
-
-using PointcloudXYZIRADT = pcl::PointCloud<PointXYZIRADT>::Ptr;
-
-POINT_CLOUD_REGISTER_POINT_STRUCT(
-  PointXYZIR,
-  (float, x, x)(float, y, y)(float, z, z)(float, intensity, intensity)(std::uint16_t, ring, ring))
-
-POINT_CLOUD_REGISTER_POINT_STRUCT(
-  PointXYZIRADT,
-  (float, x, x)(float, y, y)(float, z, z)(float, intensity, intensity)(std::uint16_t, ring, ring)(
-    float, azimuth, azimuth)(float, distance, distance)(double, time_stamp, time_stamp))
