--- hesai_pandar/include/hesai_pandar/socket_input.h
@@ -1,34 +1,0 @@
-/******************************************************************************
- * Copyright 2019 The Hesai Technology Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *****************************************************************************/
-
-#pragma once
-
-#include <netinet/in.h>
-#include <string>
-#include "hesai_pandar/input.h"
-
-class SocketInput : public Input
-{
-public:
-  SocketInput(uint16_t port, uint16_t gpsPort);
-  ~SocketInput();
-  int getPacket(pandar_msgs::PandarPacket * pkt) override;
-
-private:
-  int socketForLidar;
-  int socketForGPS;
-  int socketNumber;
-};
