--- hesai_pandar/launch/hesai_pandar.launch
@@ -1,27 +1,0 @@
-<launch>
-  <arg name="pcap"  default=""/>
-  <arg name="device_ip" default="192.168.1.201"/>
-  <arg name="lidar_port"  default="2368"/>
-  <arg name="gps_port"  default="10110"/>
-  <arg name="scan_phase"  default="0"/>
-  <arg name="model" default="Pandar40P"/>
-  <arg name="frame_id" default="pandar"/>
-  <arg name="timestamp_type" default=""/>
-  <arg name="launch_driver" default="true"/>
-  <arg name="calibration"  default="$(find hesai_pandar)/config/40p.csv"/>
-
-  <node pkg="hesai_pandar" name="hesai_pandar" type="hesai_pandar" output="screen" >
-    <remap from="pandar_points" to="pointcloud_raw" />
-    <remap from="pandar_points_ex" to="pointcloud_raw_ex" />
-    <param name="pcap" type="string" value="$(arg pcap)"/>
-    <param name="device_ip" type="string" value="$(arg device_ip)"/>
-    <param name="lidar_port"  type="int" value="$(arg lidar_port)"/>
-    <param name="gps_port"  type="int" value="$(arg gps_port)"/>
-    <param name="scan_phase"  type="double" value="$(arg scan_phase)"/>
-    <param name="model"  type="string" value="$(arg model)"/>
-    <param name="frame_id"  type="string" value="$(arg frame_id)"/>
-    <param name="calibration"  type="string" value="$(arg calibration)"/>
-    <param name="publish_packet" value="$(arg launch_driver)"/>
-    <!-- <param name="timestamp_type"  type="string" value="$(arg timestamp_type)"/> -->
-  </node>
-</launch>
