--- hesai_pandar/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>hesai_pandar</name>
-  <version>0.0.0</version>
-  <description>
-    The hesai_lidar package includes the driver for the Heisai PandarQT/Pandar64/Pandar40P/Pandar20A/Pandar20B/Pandar40M LiDAR sensor.
-  </description>
-  <maintainer email="shinnosuke.hirakawa@tier.jp">hirakawa</maintainer>
-  <license>Apache 2</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-
-  <depend>roscpp</depend>
-  <depend>roslib</depend>
-  <depend>std_msgs</depend>
-  <depend>sensor_msgs</depend>
-  <!-- <depend>libpcl-all-dev</depend> -->
-  <depend>image_transport</depend>
-  <depend>pcl_conversions</depend>
-  <depend>pcl_ros</depend>
-  <depend>pandar_msgs</depend>
-</package>
