+++ pandar_pointcloud/include/pandar_pointcloud/decoder/pandar_qt_decoder.hpp
@@ -1,0 +1,48 @@
+#pragma once
+
+#include <array>
+#include "pandar_pointcloud/calibration.hpp"
+#include "packet_decoder.hpp"
+#include "pandar_qt.hpp"
+
+namespace pandar_pointcloud
+{
+namespace pandar_qt
+{
+class PandarQTDecoder : public PacketDecoder
+{
+public:
+  enum class ReturnMode : int8_t {
+    DUAL,
+    FIRST,
+    LAST,
+  };
+
+  PandarQTDecoder(Calibration & calibration, float scan_phase = 0.0f);
+  void unpack(const pandar_msgs::PandarPacket & raw_packet) override;
+  bool hasScanned() override;
+  PointcloudXYZIRADT getPointcloud() override;
+
+private:
+  bool parsePacket(const pandar_msgs::PandarPacket & raw_packet);
+  PointcloudXYZIRADT convert(const int block_id);
+
+  std::array<float, UNIT_NUM> elev_angle_;
+  std::array<float, UNIT_NUM> azimuth_offset_;
+
+  std::array<float, UNIT_NUM> firing_offset_;
+  std::array<float, BLOCK_NUM> block_offset_single_;
+  std::array<float, BLOCK_NUM> block_offset_dual_;
+
+  Packet packet_;
+
+  PointcloudXYZIRADT scan_pc_;
+  PointcloudXYZIRADT overflow_pc_;
+
+  uint16_t scan_phase_;
+  int last_phase_;
+  bool has_scanned_;
+};
+
+}  // namespace pandar_qt
+}  // namespace pandar_pointcloud
