+++ pandar_pointcloud/nodelet_pandar_pointcloud.xml
@@ -1,0 +1,11 @@
+<class_libraries>
+  <library path="lib/libpandar_cloud_nodelet">
+    <class name="pandar_pointcloud/CloudNodelet"
+          type="pandar_pointcloud::CloudNodelet"
+          base_class_type="nodelet::Nodelet">
+      <description>
+        pandar pointcloud pulisher.
+      </description>
+    </class>
+  </library>
+</class_libraries>
