+++ pandar_pointcloud/package.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>pandar_pointcloud</name>
+  <version>0.0.0</version>
+  <description>
+    Point cloud converter for Hesai LiDARSs.
+  </description>
+  <maintainer email="shinnosuke.hirakawa@tier4.jp">hirakawa</maintainer>
+  <license>Apache 2</license>
+
+  <buildtool_depend>catkin</buildtool_depend>
+
+  <depend>roscpp</depend>
+  <depend>roslib</depend>
+  <depend>nodelet</depend>
+  <depend>pluginlib</depend>
+  <depend>pcl_conversions</depend>
+  <depend>pcl_ros</depend>
+  <depend>sensor_msgs</depend>
+  <depend>pandar_msgs</depend>
+  <depend>pandar_driver</depend>
+  
+  <export>
+    <nodelet plugin="${prefix}/nodelet_pandar_pointcloud.xml"/>
+  </export>
+</package>
